requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
djangorestframework==3.11.0
idna==2.9
joblib==0.15.1
numpy==1.18.4
pandas==1.0.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.0.0
urllib3==1.25.9
